+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaMirrorMaker2AssemblyOperator.java
@@ -4,6 +4,8 @@
  */
 package io.strimzi.operator.cluster.operator.assembly;
 
+import java.io.Serializable;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -376,6 +378,7 @@ public class KafkaMirrorMaker2AssemblyOperator extends AbstractConnectOperator<K
                 .setHandler(result -> {
                     if (result.succeeded()) {
                         mirrorMaker2Status.getConnectors().add(result.result());
+                        mirrorMaker2Status.getConnectors().sort(new ConnectorsComparatorByName());
                     } else {
                         maybeUpdateMirrorMaker2Status(reconciliation, mirrorMaker2, result.cause());
                     }
@@ -392,6 +395,21 @@ public class KafkaMirrorMaker2AssemblyOperator extends AbstractConnectOperator<K
             (mirror1, status2) -> {
                 return new KafkaMirrorMaker2Builder(mirror1).withStatus(status2).build();
             });
+    }
+
+    /**
+     * This comparator compares two maps where connectors' configurations are stored.
+     * The comparison is done by using only one property - 'name'
+     */
+    static class ConnectorsComparatorByName implements Comparator<Map<String, Object>>, Serializable {
+        private static final long serialVersionUID = 1L;
+
+        @Override
+        public int compare(Map<String, Object> m1, Map<String, Object> m2) {
+            String name1 = m1.get("name") == null ? "" : m1.get("name").toString();
+            String name2 = m2.get("name") == null ? "" : m2.get("name").toString();
+            return name1.compareTo(name2);
+        }
     }
 
 }