+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add support for Kafka 2.4.1
 * Add support for Kafka 2.5.0
 * Remove TLS sidecars from ZooKeeper pods, using native ZooKeeper TLS support instead
+* Add metrics for Topic Operator
 
 ## 0.17.0